+++ src/OpenTK.NT/Native/User32/Structs/RawInputDevice.cs
@@ -63,7 +63,7 @@ namespace OpenTK.NT.Native
             usUsagePage = usagePage;
         }
 
-        public static readonly int SizeInBytes = Marshal.SizeOf<RawInputDevice>();
+        public static readonly uint SizeInBytes = (uint)Marshal.SizeOf<RawInputDevice>();
 
         public override string ToString()
             => $"{usUsagePage}/{usUsage}, flags: {dwFlags}, window: {hwndTarget}";
